settings: 拆出 BackupSyncSettingsController（双 sub-controller 第 1 步）
将 SettingsController 中备份与同步相关的 ~40 个方法、6 个状态字段、备份的
DOM 事件绑定整体拆到 js/controller/settings/backup-sync-settings-controller.js；
共享纯函数 helper 抽到 settings/settings-utils.js（backup 与后续 mpc sub 复用）。

- 主 class 持有 this.backupController 实例；bindEvents 委托 backupController.bindEvents()；
  loadBackupSyncSettings 委托 backupController.loadSettings()。popup-controller 公开 API
  与调用点零改动。
- 保留 normalizeBasicAuth/normalizeUcanToken（经核实 MPC 协调器 UCAN 流程仍在用）
  及 normalize*/getDefault*/isLegacy* 系列（MPC 可能用），待 MPC 拆分后统一清理。
- 主 class 3077 → 1914 行。新增 tests/settings-utils.test.mjs（7 例）。
- 已做模块加载 + 实例化冒烟、残留引用扫描；UI 行为需在 Chrome 实测（见下）。

注：MPC 设置仍在主 class，作为双 sub-controller 拆分的第 2 步后续处理。
