+++ js/controller/settings-controller.js
@@ -11,6 +11,7 @@ import {
   createUcanInvocationToken
 } from '../common/ucan-utils.js';
 import { isDeveloperFeatureEnabled } from '../config/index.js';
+import { BackupSyncSettingsController } from './settings/backup-sync-settings-controller.js';
 
 const LEGACY_DEFAULT_APP_ID = 'yeying-wallet';
 const DEFAULT_UCAN_ACTION = 'write';
@@ -42,13 +43,8 @@ export class SettingsController {
     this.cachedSites = [];
     this.activeSiteDetail = null;
     this.resetConfirmKeyword = 'RESET';
-    this.syncSettings = null;
     this.mpcSettings = null;
-    this.syncLogs = [];
-    this.syncLogQuery = '';
-    this.syncLogFiltered = [];
-    this.syncLogPageSize = 30;
-    this.syncLogVisibleCount = 0;
+    this.backupController = new BackupSyncSettingsController({ wallet, transaction, requestPassword });
     this.mpcLogs = [];
     this.mpcLogQuery = '';
     this.mpcLogFiltered = [];
@@ -65,104 +61,7 @@ export class SettingsController {
   }
 
   bindEvents() {
-    const backupSyncEnabledToggle = document.getElementById('backupSyncEnabledToggle');
-    if (backupSyncEnabledToggle) {
-      backupSyncEnabledToggle.addEventListener('change', async () => {
-        await this.handleBackupSyncToggle(backupSyncEnabledToggle.checked);
-      });
-    }
-
-    const backupSyncEndpointInput = document.getElementById('backupSyncEndpointInput');
-    if (backupSyncEndpointInput) {
-      backupSyncEndpointInput.addEventListener('blur', async () => {
-        await this.handleBackupSyncEndpointUpdate(backupSyncEndpointInput.value);
-      });
-    }
-
-    const backupSyncAuthModeSelect = document.getElementById('backupSyncAuthModeSelect');
-    if (backupSyncAuthModeSelect) {
-      backupSyncAuthModeSelect.addEventListener('change', async () => {
-        const mode = backupSyncAuthModeSelect.value;
-        await this.handleBackupSyncAuthModeChange(mode);
-      });
-    }
-
-    const backupSyncSiweLoginBtn = document.getElementById('backupSyncSiweLoginBtn');
-    if (backupSyncSiweLoginBtn) {
-      backupSyncSiweLoginBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncSiweLogin();
-      });
-    }
-
-    const backupSyncSiweRefreshBtn = document.getElementById('backupSyncSiweRefreshBtn');
-    if (backupSyncSiweRefreshBtn) {
-      backupSyncSiweRefreshBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncSiweRefresh();
-      });
-    }
-
-    const backupSyncUcanGenerateBtn = document.getElementById('backupSyncUcanGenerateBtn');
-    if (backupSyncUcanGenerateBtn) {
-      backupSyncUcanGenerateBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncUcanGenerate();
-      });
-    }
-
-    const backupSyncBasicSaveBtn = document.getElementById('backupSyncBasicSaveBtn');
-    if (backupSyncBasicSaveBtn) {
-      backupSyncBasicSaveBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncBasicSave();
-      });
-    }
-
-    const backupSyncNowBtn = document.getElementById('backupSyncNowBtn');
-    if (backupSyncNowBtn) {
-      backupSyncNowBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncNow();
-      });
-    }
-
-    const backupSyncClearRemoteBtn = document.getElementById('backupSyncClearRemoteBtn');
-    if (backupSyncClearRemoteBtn) {
-      backupSyncClearRemoteBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncClearRemote();
-      });
-    }
-
-    const backupSyncClearLogsBtn = document.getElementById('backupSyncClearLogsBtn');
-    if (backupSyncClearLogsBtn) {
-      backupSyncClearLogsBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncClearLogs();
-      });
-    }
-
-    const backupSyncLogRetentionSaveBtn = document.getElementById('backupSyncLogRetentionSaveBtn');
-    if (backupSyncLogRetentionSaveBtn) {
-      backupSyncLogRetentionSaveBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncLogRetentionSave();
-      });
-    }
-
-    const backupSyncAuditBtn = document.getElementById('backupSyncAuditBtn');
-    if (backupSyncAuditBtn) {
-      backupSyncAuditBtn.addEventListener('click', async () => {
-        await this.openBackupSyncLogsPage();
-      });
-    }
-
-    const backupSyncLogsSearchInput = document.getElementById('backupSyncLogsSearchInput');
-    if (backupSyncLogsSearchInput) {
-      backupSyncLogsSearchInput.addEventListener('input', () => {
-        this.applyBackupSyncLogsFilter(backupSyncLogsSearchInput.value);
-      });
-    }
-
-    const backupSyncLogsList = document.getElementById('backupSyncLogsList');
-    if (backupSyncLogsList) {
-      backupSyncLogsList.addEventListener('scroll', () => {
-        this.handleBackupSyncLogsScroll();
-      });
-    }
+    this.backupController.bindEvents();
 
     const mpcAuditBtn = document.getElementById('mpcAuditBtn');
     if (mpcAuditBtn) {
@@ -374,25 +273,6 @@ export class SettingsController {
       }
     }
 
-    const backupSyncSimulateBtn = document.getElementById('backupSyncSimulateConflictBtn');
-    if (backupSyncSimulateBtn) {
-      backupSyncSimulateBtn.addEventListener('click', async () => {
-        await this.handleBackupSyncSimulateConflict();
-      });
-    }
-
-    const conflictsList = document.getElementById('backupSyncConflictsList');
-    if (conflictsList) {
-      conflictsList.addEventListener('click', async (event) => {
-        const actionBtn = event.target.closest('[data-conflict-action]');
-        if (!actionBtn) return;
-        const conflictId = actionBtn.dataset.conflictId;
-        const action = actionBtn.dataset.conflictAction;
-        if (!conflictId || !action) return;
-        await this.handleResolveBackupSyncConflict(conflictId, action);
-      });
-    }
-
     const changePasswordBtn = document.getElementById('changePasswordBtn');
     if (changePasswordBtn) {
       changePasswordBtn.addEventListener('click', () => {
@@ -558,13 +438,7 @@ export class SettingsController {
   }
 
   async loadBackupSyncSettings() {
-    try {
-      const settings = await this.wallet.getBackupSyncSettings();
-      this.syncSettings = settings || {};
-      this.renderBackupSyncSettings(settings);
-    } catch (error) {
-      console.error('[SettingsController] 获取 Backup & Sync 设置失败:', error);
-    }
+    return this.backupController.loadSettings();
   }
 
   async loadMpcSettings() {
@@ -576,53 +450,6 @@ export class SettingsController {
     } catch (error) {
       console.error('[SettingsController] 获取 MPC 设置失败:', error);
     }
-  }
-
-  renderBackupSyncSettings(settings = {}) {
-    const enabledToggle = document.getElementById('backupSyncEnabledToggle');
-    const endpointInput = document.getElementById('backupSyncEndpointInput');
-    const authModeSelect = document.getElementById('backupSyncAuthModeSelect');
-    const ucanResourceInput = document.getElementById('backupSyncUcanResourceInput');
-    const ucanActionInput = document.getElementById('backupSyncUcanActionInput');
-    const ucanAudienceInput = document.getElementById('backupSyncUcanAudienceInput');
-    const ucanTtlInput = document.getElementById('backupSyncUcanTtlInput');
-    const basicInput = document.getElementById('backupSyncBasicInput');
-    const tokenStatus = document.getElementById('backupSyncTokenStatus');
-    const lastStatus = document.getElementById('backupSyncLastStatus');
-
-    if (enabledToggle) enabledToggle.checked = Boolean(settings.enabled);
-    if (endpointInput) endpointInput.value = settings.endpoint || 'https://webdav.yeying.pub/dav';
-    if (authModeSelect) authModeSelect.value = settings.authMode || 'ucan';
-    if (ucanResourceInput) {
-      const normalizedResource = this.normalizeUcanResource(settings.ucanResource || '');
-      ucanResourceInput.value = normalizedResource;
-      ucanResourceInput.readOnly = true;
-    }
-    if (ucanActionInput) {
-      const normalizedResource = this.normalizeUcanResource(settings.ucanResource || '');
-      ucanActionInput.value = this.normalizeUcanAction(settings.ucanAction || '', normalizedResource);
-      ucanActionInput.readOnly = true;
-    }
-    if (ucanAudienceInput) {
-      ucanAudienceInput.value = settings.ucanAudience || this.deriveUcanAudience(settings.endpoint || '');
-      ucanAudienceInput.readOnly = true;
-    }
-    if (ucanTtlInput && !ucanTtlInput.value) {
-      ucanTtlInput.value = '24';
-    }
-    if (basicInput) basicInput.value = settings.basicAuth || '';
-
-    this.updateBackupSyncAuthPanel(settings.authMode || 'ucan');
-    this.updateBackupSyncTokenStatus(settings, tokenStatus);
-    this.updateBackupSyncLastStatus(settings, lastStatus);
-    this.updateBackupSyncEnabledState(Boolean(settings.enabled));
-    window.refreshWalletSelects?.();
-
-    this.renderBackupSyncAccountAddress();
-    this.renderBackupSyncConflicts(settings.conflicts || []);
-    this.renderBackupSyncLogRetention(settings);
-    this.updateBackupSyncLogs(settings.logs || []);
-    this.toggleBackupSyncDebug();
   }
 
   renderMpcSettings(settings = {}) {
@@ -1331,273 +1158,6 @@ export class SettingsController {
     }
   }
 
-  updateBackupSyncAuthPanel(mode) {
-    const siwePanel = document.getElementById('backupSyncSiwePanel');
-    const ucanPanel = document.getElementById('backupSyncUcanPanel');
-    const basicPanel = document.getElementById('backupSyncBasicPanel');
-
-    if (siwePanel) siwePanel.classList.toggle('hidden', mode !== 'siwe');
-    if (ucanPanel) ucanPanel.classList.toggle('hidden', mode !== 'ucan');
-    if (basicPanel) basicPanel.classList.toggle('hidden', mode !== 'basic');
-  }
-
-  updateBackupSyncEnabledState(enabled) {
-    const endpointInput = document.getElementById('backupSyncEndpointInput');
-    const authModeSelect = document.getElementById('backupSyncAuthModeSelect');
-    const nowBtn = document.getElementById('backupSyncNowBtn');
-
-    if (endpointInput) endpointInput.disabled = !enabled;
-    if (authModeSelect) authModeSelect.disabled = !enabled;
-    if (nowBtn) nowBtn.disabled = !enabled;
-
-    const panelControls = document.querySelectorAll(
-      '#backupSyncSiwePanel button, #backupSyncUcanPanel button, #backupSyncUcanPanel input, #backupSyncUcanPanel textarea, #backupSyncBasicPanel button, #backupSyncBasicPanel input'
-    );
-    panelControls.forEach(el => {
-      el.disabled = !enabled;
-    });
-  }
-
-  updateBackupSyncTokenStatus(settings, element) {
-    if (!element) return;
-    const token = settings?.authToken || '';
-    const expiresAt = settings?.authTokenExpiresAt;
-    if (!token) {
-      element.textContent = '未登录';
-      return;
-    }
-    const expiresText = expiresAt ? formatLocaleDateTime(expiresAt) : '未知';
-    element.textContent = `已登录 · 过期时间 ${expiresText}`;
-  }
-
-  updateBackupSyncLastStatus(settings, element) {
-    if (!element) return;
-    const pullText = settings?.lastPullAt ? formatLocaleDateTime(settings.lastPullAt) : '-';
-    const pushText = settings?.lastPushAt ? formatLocaleDateTime(settings.lastPushAt) : '-';
-    element.textContent = `最近拉取: ${pullText} · 最近推送: ${pushText}`;
-  }
-
-  renderBackupSyncConflicts(conflicts = []) {
-    const container = document.getElementById('backupSyncConflictsList');
-    if (!container) return;
-    const list = Array.isArray(conflicts) ? conflicts : [];
-
-    if (list.length === 0) {
-      container.innerHTML = '<div class="empty-message">暂无冲突</div>';
-      return;
-    }
-
-    container.innerHTML = list.map(conflict => {
-      const title = conflict.type === 'contact'
-        ? `联系人 ${conflict.address ? shortenAddress(conflict.address) : ''}`
-        : `账户 #${conflict.index ?? '-'}`;
-      const localName = escapeHtml(conflict.localName || '');
-      const remoteName = escapeHtml(conflict.remoteName || '');
-      const timeText = conflict.timestamp ? formatLocaleDateTime(conflict.timestamp) : '-';
-
-      return `
-        <div class="sync-conflict-item">
-          <div class="sync-conflict-info">
-            <div class="sync-conflict-title">${escapeHtml(title)}</div>
-            <div class="sync-conflict-meta">时间: ${escapeHtml(timeText)}</div>
-            <div class="sync-conflict-names">
-              <span class="sync-conflict-local">本地: ${localName || '-'}</span>
-              <span class="sync-conflict-remote">远端: ${remoteName || '-'}</span>
-            </div>
-          </div>
-          <div class="sync-conflict-actions">
-            <button class="btn btn-secondary btn-small" data-conflict-action="local" data-conflict-id="${escapeHtml(conflict.id)}">用本地</button>
-            <button class="btn btn-primary btn-small" data-conflict-action="remote" data-conflict-id="${escapeHtml(conflict.id)}">用远端</button>
-          </div>
-        </div>
-      `;
-    }).join('');
-  }
-
-  async openBackupSyncLogsPage() {
-    showPage('backupSyncLogsPage');
-    await this.loadBackupSyncLogs();
-  }
-
-  async loadBackupSyncLogs() {
-    try {
-      const settings = await this.wallet.getBackupSyncSettings();
-      this.syncSettings = settings || this.syncSettings || {};
-      this.renderBackupSyncLogRetention(settings);
-      this.updateBackupSyncLogs(settings?.logs || []);
-    } catch (error) {
-      console.error('[SettingsController] 加载同步日志失败:', error);
-      this.updateBackupSyncLogs([]);
-    }
-  }
-
-  renderBackupSyncLogRetention(settings = {}) {
-    const maxInput = document.getElementById('backupSyncLogMaxInput');
-    const daysInput = document.getElementById('backupSyncLogRetentionInput');
-    const maxCount = this.normalizeLogMaxCount(settings.logMaxCount ?? DEFAULT_LOG_MAX_COUNT);
-    const days = this.normalizeLogRetentionDays(settings.logRetentionDays ?? DEFAULT_LOG_RETENTION_DAYS);
-
-    if (maxInput) {
-      maxInput.value = String(maxCount);
-    }
-    if (daysInput) {
-      daysInput.value = String(days);
-    }
-  }
-
-  updateBackupSyncLogs(logs = []) {
-    this.syncLogs = Array.isArray(logs) ? logs : [];
-    this.updateBackupSyncLogsSummary();
-    this.applyBackupSyncLogsFilter();
-  }
-
-  applyBackupSyncLogsFilter(keyword) {
-    const inputValue = arguments.length > 0
-      ? keyword
-      : (document.getElementById('backupSyncLogsSearchInput')?.value || '');
-    const normalized = String(inputValue || '').trim().toLowerCase();
-    this.syncLogQuery = normalized;
-    const source = Array.isArray(this.syncLogs) ? this.syncLogs : [];
-    this.syncLogFiltered = normalized
-      ? source.filter(entry => this.matchBackupSyncLog(entry, normalized))
-      : [...source];
-    this.syncLogVisibleCount = Math.min(this.syncLogPageSize, this.syncLogFiltered.length);
-    this.updateBackupSyncLogsSummary();
-    this.renderBackupSyncLogsList(false);
-
-    const container = document.getElementById('backupSyncLogsList');
-    if (container) {
-      container.scrollTop = 0;
-    }
-  }
-
-  matchBackupSyncLog(entry, keyword) {
-    if (!keyword) return true;
-    const timeText = entry?.time ? formatLocaleDateTime(entry.time) : '';
-    const fields = [
-      entry?.message,
-      entry?.action,
-      entry?.reason,
-      entry?.level,
-      entry?.id,
-      timeText
-    ].filter(Boolean);
-    const haystack = fields.join(' ').toLowerCase();
-    return haystack.includes(keyword);
-  }
-
-  handleBackupSyncLogsScroll() {
-    const container = document.getElementById('backupSyncLogsList');
-    if (!container) return;
-    if (this.syncLogVisibleCount >= this.syncLogFiltered.length) return;
-    const threshold = 24;
-    if (container.scrollTop + container.clientHeight >= container.scrollHeight - threshold) {
-      this.syncLogVisibleCount = Math.min(
-        this.syncLogFiltered.length,
-        this.syncLogVisibleCount + this.syncLogPageSize
-      );
-      this.renderBackupSyncLogsList(true);
-    }
-  }
-
-  renderBackupSyncLogsList(preserveScroll = false) {
-    const container = document.getElementById('backupSyncLogsList');
-    if (!container) return;
-    const list = Array.isArray(this.syncLogFiltered) ? this.syncLogFiltered : [];
-    const total = list.length;
-    const baseCount = this.syncLogVisibleCount || this.syncLogPageSize;
-    const visibleCount = Math.min(baseCount, total);
-    this.syncLogVisibleCount = visibleCount;
-    const entries = list.slice(0, visibleCount);
-    const scrollTop = preserveScroll ? container.scrollTop : 0;
-
-    if (entries.length === 0) {
-      const emptyText = this.syncLogQuery ? '没有匹配的日志' : '暂无日志';
-      container.innerHTML = `<div class="empty-message">${emptyText}</div>`;
-      this.updateBackupSyncLogsFooter(total, visibleCount);
-      return;
-    }
-
-    container.innerHTML = entries.map(entry => {
-      const timeText = entry?.time ? formatLocaleDateTime(entry.time) : '-';
-      const message = escapeHtml(entry?.message || '-');
-      const level = String(entry?.level || 'info').toLowerCase();
-      const levelClass = level === 'error' ? 'error' : level === 'warn' ? 'warn' : 'info';
-      const levelLabel = levelClass === 'error' ? '错误' : levelClass === 'warn' ? '警告' : '信息';
-      const reasonLabel = this.formatBackupSyncReason(entry?.reason || '');
-      const durationText = Number.isFinite(entry?.durationMs)
-        ? `${Math.max(0, Math.round(entry.durationMs / 1000))}s`
-        : '';
-      const actionLabel = entry?.action ? `动作 ${entry.action}` : '';
-
-      return `
-        <div class="sync-activity-item">
-          <div class="sync-activity-time">${escapeHtml(timeText)}</div>
-          <div class="sync-activity-main">
-            <div class="sync-activity-message">${message}</div>
-            <div class="sync-activity-meta">
-              <span class="sync-activity-tag level-${levelClass}">${escapeHtml(levelLabel)}</span>
-              ${actionLabel ? `<span class="sync-activity-tag">${escapeHtml(actionLabel)}</span>` : ''}
-              ${reasonLabel ? `<span class="sync-activity-tag">${escapeHtml(reasonLabel)}</span>` : ''}
-              ${durationText ? `<span class="sync-activity-tag">耗时 ${escapeHtml(durationText)}</span>` : ''}
-            </div>
-          </div>
-        </div>
-      `;
-    }).join('');
-
-    if (preserveScroll) {
-      container.scrollTop = scrollTop;
-    }
-
-    this.updateBackupSyncLogsFooter(total, visibleCount);
-  }
-
-  updateBackupSyncLogsFooter(total, visibleCount) {
-    const footer = document.getElementById('backupSyncLogsFooter');
-    if (!footer) return;
-    if (!total) {
-      footer.classList.add('hidden');
-      footer.textContent = '';
-      return;
-    }
-    footer.classList.remove('hidden');
-    if (visibleCount < total) {
-      footer.textContent = `已加载 ${visibleCount} / ${total}，向下滚动加载更多`;
-    } else {
-      footer.textContent = `已加载全部 ${total} 条日志`;
-    }
-  }
-
-  updateBackupSyncLogsSummary() {
-    const lastPullEl = document.getElementById('backupSyncLogsLastPull');
-    const lastPushEl = document.getElementById('backupSyncLogsLastPush');
-    const totalEl = document.getElementById('backupSyncLogsTotal');
-    const matchEl = document.getElementById('backupSyncLogsMatch');
-
-    if (lastPullEl) {
-      const pullText = this.syncSettings?.lastPullAt
-        ? formatLocaleDateTime(this.syncSettings.lastPullAt)
-        : '-';
-      lastPullEl.textContent = pullText;
-    }
-
-    if (lastPushEl) {
-      const pushText = this.syncSettings?.lastPushAt
-        ? formatLocaleDateTime(this.syncSettings.lastPushAt)
-        : '-';
-      lastPushEl.textContent = pushText;
-    }
-
-    if (totalEl) {
-      totalEl.textContent = String(Array.isArray(this.syncLogs) ? this.syncLogs.length : 0);
-    }
-
-    if (matchEl) {
-      matchEl.textContent = String(Array.isArray(this.syncLogFiltered) ? this.syncLogFiltered.length : 0);
-    }
-  }
-
   async openMpcLogsPage() {
     showPage('mpcLogsPage');
     await this.loadMpcLogs();
@@ -1954,561 +1514,8 @@ export class SettingsController {
     }
   }
 
-  formatBackupSyncReason(reason = '') {
-    if (!reason) return '';
-    if (reason.startsWith('debounced:')) {
-      const detail = reason.slice('debounced:'.length);
-      return detail ? `本地变更(${detail})` : '本地变更';
-    }
-    switch (reason) {
-      case 'manual':
-        return '手动';
-      case 'auto':
-        return '自动';
-      case 'unlock':
-        return '解锁';
-      case 'lock':
-        return '锁定';
-      default:
-        return reason;
-    }
-  }
-
-  toggleBackupSyncDebug() {
-    const debugRow = document.getElementById('backupSyncDebugRow');
-    if (!debugRow) return;
-    const enabled = isDeveloperFeatureEnabled('ENABLE_DEBUG_MODE');
-    debugRow.classList.toggle('hidden', !enabled);
-  }
-
-  async handleResolveBackupSyncConflict(conflictId, action) {
-    try {
-      const result = await this.wallet.resolveBackupSyncConflict({ id: conflictId, action });
-      if (!result?.success) {
-        throw new Error(result?.error || '处理失败');
-      }
-      await this.loadBackupSyncSettings();
-      showSuccess('已处理冲突');
-    } catch (error) {
-      console.error('[SettingsController] 处理同步冲突失败:', error);
-      await this.logBackupSyncError('conflict-resolve-error', error?.message || '处理冲突失败');
-      showError('处理失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncSimulateConflict() {
-    try {
-      if (!isDeveloperFeatureEnabled('ENABLE_DEBUG_MODE')) {
-        showError('调试模式未开启');
-        return;
-      }
-
-      const current = await this.wallet.getCurrentAccount();
-      if (!current?.id) {
-        showError('未找到当前账户');
-        return;
-      }
-
-      const contacts = await this.wallet.getContacts();
-      const timestamp = Date.now();
-      const conflicts = Array.isArray(this.syncSettings?.conflicts)
-        ? [...this.syncSettings.conflicts]
-        : [];
-
-      conflicts.push({
-        id: `account:${current.id}:${timestamp}`,
-        type: 'account',
-        accountId: current.id,
-        walletId: current.walletId || '',
-        index: Number.isFinite(current.index) ? current.index : 0,
-        localName: current.name || '',
-        remoteName: `${current.name || 'Account'} (Remote)`,
-        timestamp
-      });
-
-      if (Array.isArray(contacts) && contacts.length > 0) {
-        const contact = contacts[0];
-        conflicts.push({
-          id: `contact:${contact.id}:${timestamp}`,
-          type: 'contact',
-          contactId: contact.id,
-          address: contact.address,
-          localName: contact.name || '',
-          localNote: contact.note || '',
-          remoteName: `${contact.name || 'Contact'} (Remote)`,
-          remoteNote: `${contact.note || ''}`.trim(),
-          timestamp
-        });
-      }
-
-      const result = await this.wallet.updateBackupSyncSettings({ conflicts });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-      showSuccess('已生成测试冲突');
-    } catch (error) {
-      console.error('[SettingsController] 生成测试冲突失败:', error);
-      showError('生成失败: ' + error.message);
-    }
-  }
-
   deriveUcanAudience(endpoint) {
     return deriveUcanAudience(endpoint);
-  }
-
-  async renderBackupSyncAccountAddress() {
-    const addressEl = document.getElementById('backupSyncAccountAddress');
-    if (!addressEl) return;
-    try {
-      const account = await this.wallet.getCurrentAccount();
-      addressEl.textContent = account?.address ? shortenAddress(account.address) : '-';
-      addressEl.title = account?.address || '';
-    } catch (error) {
-      addressEl.textContent = '-';
-    }
-  }
-
-  async handleBackupSyncToggle(enabled) {
-    try {
-      const result = await this.wallet.updateBackupSyncSettings({ enabled });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-      showSuccess(enabled ? '已启用备份与同步' : '已关闭备份与同步');
-    } catch (error) {
-      console.error('[SettingsController] 更新 Backup & Sync 开关失败:', error);
-      showError('更新失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncEndpointUpdate(endpoint) {
-    const trimmed = String(endpoint || '').trim();
-    if (!trimmed) {
-      showError('请输入 WebDAV 地址');
-      return;
-    }
-    try {
-      new URL(trimmed);
-    } catch {
-      showError('WebDAV 地址格式不正确');
-      return;
-    }
-    try {
-      const resolved = await this.detectBackupSyncEndpoint(trimmed);
-      const result = await this.wallet.updateBackupSyncSettings({ endpoint: resolved });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-      if (resolved !== trimmed) {
-        showSuccess(`WebDAV 地址已保存（自动识别为 ${resolved}）`);
-      } else {
-        showSuccess('WebDAV 地址已保存');
-      }
-    } catch (error) {
-      console.error('[SettingsController] 更新 WebDAV 地址失败:', error);
-      await this.logBackupSyncError('endpoint-update-error', error?.message || 'WebDAV 地址保存失败');
-      showError('保存失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncAuthModeChange(mode) {
-    try {
-      const result = await this.wallet.updateBackupSyncSettings({ authMode: mode });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-    } catch (error) {
-      console.error('[SettingsController] 更新认证方式失败:', error);
-      showError('更新失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncSiweLogin() {
-    try {
-      const endpoint = (document.getElementById('backupSyncEndpointInput')?.value || '').trim();
-      if (!endpoint) {
-        showError('请输入 WebDAV 地址');
-        return;
-      }
-      try {
-        new URL(endpoint);
-      } catch {
-        showError('WebDAV 地址格式不正确');
-        return;
-      }
-
-      const account = await this.wallet.getCurrentAccount();
-      if (!account?.address) {
-        showError('未找到当前账户');
-        return;
-      }
-
-      if (!this.transaction) {
-        showError('签名模块未初始化');
-        return;
-      }
-
-      showWaiting();
-      const challenge = await this.fetchSiweChallenge(endpoint, account.address);
-      if (!challenge) {
-        showError('无法获取挑战信息');
-        await this.logBackupSyncError('siwe-login-error', '无法获取挑战信息');
-        return;
-      }
-      hideWaiting();
-
-      const password = await this.requestPassword?.();
-      if (!password) {
-        return;
-      }
-
-      showWaiting();
-      const signature = await this.transaction.signMessage(challenge, password);
-      const verifyResult = await this.fetchSiweVerify(endpoint, account.address, signature);
-      const token = verifyResult?.token;
-      const expiresAt = verifyResult?.expiresAt || null;
-
-      if (!token) {
-        showError('登录失败：未返回 Token');
-        await this.logBackupSyncError('siwe-login-error', '未返回 Token');
-        return;
-      }
-
-      const result = await this.wallet.updateBackupSyncSettings({
-        authMode: 'siwe',
-        authToken: token,
-        authTokenExpiresAt: expiresAt
-      });
-
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-
-      showSuccess('SIWE 登录成功');
-    } catch (error) {
-      console.error('[SettingsController] SIWE 登录失败:', error);
-      await this.logBackupSyncError('siwe-login-error', error?.message || 'SIWE 登录失败');
-      showError('登录失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncSiweRefresh() {
-    try {
-      const endpoint = (document.getElementById('backupSyncEndpointInput')?.value || '').trim();
-      if (!endpoint) {
-        showError('请输入 WebDAV 地址');
-        return;
-      }
-      try {
-        new URL(endpoint);
-      } catch {
-        showError('WebDAV 地址格式不正确');
-        return;
-      }
-
-      showWaiting();
-      const refreshResult = await this.fetchSiweRefresh(endpoint);
-      const token = refreshResult?.token;
-      const expiresAt = refreshResult?.expiresAt || null;
-
-      if (!token) {
-        showError('刷新失败：未返回 Token');
-        await this.logBackupSyncError('siwe-refresh-error', '未返回 Token');
-        return;
-      }
-
-      const result = await this.wallet.updateBackupSyncSettings({
-        authMode: 'siwe',
-        authToken: token,
-        authTokenExpiresAt: expiresAt
-      });
-
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-
-      showSuccess('Token 已刷新');
-    } catch (error) {
-      console.error('[SettingsController] 刷新 Token 失败:', error);
-      await this.logBackupSyncError('siwe-refresh-error', error?.message || 'SIWE 刷新失败');
-      showError('刷新失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncUcanGenerate() {
-    try {
-      const generation = await this.generateBackupSyncUcan();
-      if (generation?.cancelled) {
-        return;
-      }
-      await this.tryBackupSyncAfterUcanAuth(generation?.settings);
-      showSuccess('UCAN 已生成');
-    } catch (error) {
-      console.error('[SettingsController] 生成 UCAN 失败:', error);
-      await this.logBackupSyncError('ucan-generate-error', error?.message || 'UCAN 生成失败');
-      showError('生成失败: ' + error.message);
-    }
-  }
-
-  async generateBackupSyncUcan() {
-    const endpoint = (document.getElementById('backupSyncEndpointInput')?.value || '').trim();
-    if (!endpoint) {
-      throw new Error('请输入 WebDAV 地址');
-    }
-    try {
-      new URL(endpoint);
-    } catch {
-      throw new Error('WebDAV 地址格式不正确');
-    }
-
-    const account = await this.wallet.getCurrentAccount();
-    if (!account?.address) {
-      throw new Error('未找到当前账户');
-    }
-
-    if (!this.transaction) {
-      throw new Error('签名模块未初始化');
-    }
-
-    const resourceInput = document.getElementById('backupSyncUcanResourceInput');
-    const actionInput = document.getElementById('backupSyncUcanActionInput');
-    const audienceInput = document.getElementById('backupSyncUcanAudienceInput');
-    const ttlInput = document.getElementById('backupSyncUcanTtlInput');
-
-    const resource = this.normalizeUcanResource(resourceInput?.value || '');
-    const action = this.normalizeUcanAction(actionInput?.value || '', resource);
-    const audience = String(audienceInput?.value || '').trim() || this.deriveUcanAudience(endpoint);
-    const ttlHours = Number(ttlInput?.value || '24');
-    const ttlMs = Number.isFinite(ttlHours) ? ttlHours * 60 * 60 * 1000 : 24 * 60 * 60 * 1000;
-
-    if (!audience) {
-      throw new Error('请填写 Audience');
-    }
-
-    const now = Date.now();
-    const expiresAt = now + ttlMs;
-    const { did, keys } = await createUcanInvocationKey();
-    const statement = {
-      aud: did,
-      cap: [{ resource, action }],
-      exp: expiresAt,
-      nbf: now
-    };
-    const message = buildSiweMessage(endpoint, account.address, statement);
-
-    const password = await this.requestPassword?.();
-    if (!password) {
-      return { cancelled: true, settings: this.syncSettings || null };
-    }
-
-    showWaiting();
-    try {
-      const signature = await this.transaction.signMessage(message, password);
-      const rootProof = {
-        type: 'siwe',
-        iss: `did:pkh:eth:${account.address.toLowerCase()}`,
-        aud: did,
-        cap: [{ resource, action }],
-        exp: expiresAt,
-        nbf: now,
-        siwe: {
-          message,
-          signature
-        }
-      };
-
-      const { token } = await createUcanInvocationToken({
-        audience,
-        capability: { resource, action },
-        proof: rootProof,
-        expiresAt,
-        notBefore: now,
-        keys,
-        did
-      });
-
-      const normalizedToken = this.normalizeUcanToken(token);
-      if (audienceInput) audienceInput.value = audience;
-      if (resourceInput) resourceInput.value = resource;
-      if (actionInput) actionInput.value = action;
-
-      const result = await this.wallet.updateBackupSyncSettings({
-        authMode: 'ucan',
-        ucanToken: normalizedToken,
-        ucanResource: resource,
-        ucanAction: action,
-        ucanAudience: audience
-      });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-      return { cancelled: false, settings: result?.settings || this.syncSettings || null };
-    } finally {
-      hideWaiting();
-    }
-  }
-
-  async handleBackupSyncBasicSave() {
-    const basicInput = document.getElementById('backupSyncBasicInput');
-    const raw = String(basicInput?.value || '').trim();
-    if (!raw) {
-      showError('请输入 Basic 凭证');
-      return;
-    }
-
-    try {
-      const header = this.normalizeBasicAuth(raw);
-      const result = await this.wallet.updateBackupSyncSettings({
-        authMode: 'basic',
-        basicAuth: header
-      });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-      showSuccess('Basic 凭证已保存');
-    } catch (error) {
-      console.error('[SettingsController] 保存 Basic 凭证失败:', error);
-      showError('保存失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncNow() {
-    try {
-      await this.ensureBackupSyncEndpoint();
-      const currentSettings = this.syncSettings || await this.wallet.getBackupSyncSettings();
-      const preflight = await this.ensureBackupSyncUcanFresh(currentSettings);
-      if (preflight?.cancelled) {
-        return;
-      }
-
-      let result;
-      showWaiting();
-      try {
-        result = await this.wallet.backupSyncNow();
-      } finally {
-        hideWaiting();
-      }
-
-      if (!result?.success) {
-        const syncError = new Error(result?.error || '同步失败');
-        if (this.shouldRetryBackupSyncWithUcan(syncError, preflight?.settings || currentSettings)) {
-          const refresh = await this.ensureBackupSyncUcanFresh(preflight?.settings || currentSettings, { force: true });
-          if (refresh?.cancelled) {
-            return;
-          }
-          showWaiting();
-          try {
-            result = await this.wallet.backupSyncNow();
-          } finally {
-            hideWaiting();
-          }
-        }
-      }
-
-      if (!result?.success) {
-        throw new Error(result?.error || '同步失败');
-      }
-      await this.loadBackupSyncSettings();
-      showSuccess('同步完成');
-    } catch (error) {
-      console.error('[SettingsController] 立即同步失败:', error);
-      showError(this.formatBackupSyncError(error, '同步失败'));
-    } finally {
-      hideWaiting();
-    }
-  }
-
-  async handleBackupSyncClearRemote() {
-    if (!confirm('确定要清除远端备份吗？此操作会删除 WebDAV 上的备份文件。')) {
-      return;
-    }
-
-    try {
-      showWaiting();
-      const result = await this.wallet.backupSyncClearRemote();
-      if (!result?.success) {
-        throw new Error(result?.error || '清理失败');
-      }
-      await this.loadBackupSyncSettings();
-      showSuccess('远端备份已清除');
-    } catch (error) {
-      console.error('[SettingsController] 清除远端备份失败:', error);
-      showError(this.formatBackupSyncError(error, '清理失败'));
-    }
-  }
-
-  async handleBackupSyncClearLogs() {
-    if (!confirm('确定要清空同步日志吗？')) {
-      return;
-    }
-
-    try {
-      const result = await this.wallet.backupSyncClearLogs();
-      if (!result?.success) {
-        throw new Error(result?.error || '清理失败');
-      }
-      await this.loadBackupSyncSettings();
-      showSuccess('同步日志已清空');
-    } catch (error) {
-      console.error('[SettingsController] 清空同步日志失败:', error);
-      showError('清理失败: ' + error.message);
-    }
-  }
-
-  async handleBackupSyncLogRetentionSave() {
-    const maxInput = document.getElementById('backupSyncLogMaxInput');
-    const daysInput = document.getElementById('backupSyncLogRetentionInput');
-    const rawMax = Number(maxInput?.value || '');
-    const rawDays = Number(daysInput?.value || '');
-
-    const maxCount = this.normalizeLogMaxCount(rawMax);
-    const days = this.normalizeLogRetentionDays(rawDays);
-
-    if (!Number.isFinite(maxCount) || maxCount <= 0) {
-      showError('最大保留条数无效');
-      return;
-    }
-    if (!Number.isFinite(days) || days < LOG_RETENTION_MIN_DAYS) {
-      showError('保留天数无效');
-      return;
-    }
-
-    try {
-      const result = await this.wallet.updateBackupSyncSettings({
-        logMaxCount: maxCount,
-        logRetentionDays: days
-      });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncLogRetention(result.settings);
-        this.updateBackupSyncLogs(result.settings.logs || []);
-      }
-      showSuccess('留存策略已保存');
-    } catch (error) {
-      console.error('[SettingsController] 保存日志留存策略失败:', error);
-      showError('保存失败: ' + error.message);
-    }
-  }
-
-  async logBackupSyncError(action, message) {
-    try {
-      if (!this.wallet?.logBackupSyncEvent) return;
-      await this.wallet.logBackupSyncEvent({
-        level: 'error',
-        action,
-        message
-      });
-    } catch (error) {
-      console.warn('[SettingsController] 写入同步错误日志失败:', error?.message || error);
-    }
   }
 
   normalizeBasicAuth(value) {
@@ -2522,47 +1529,6 @@ export class SettingsController {
 
   normalizeUcanToken(value) {
     return normalizeBearerToken(value);
-  }
-
-  getBackupSyncUcanExpiresAt(settings = {}) {
-    return getUcanExpiresAt(settings?.ucanToken || '');
-  }
-
-  isBackupSyncUcanExpiringSoon(settings = {}, thresholdMs = 5 * 60 * 1000) {
-    const expiresAt = this.getBackupSyncUcanExpiresAt(settings);
-    if (!expiresAt) return true;
-    return expiresAt <= (Date.now() + thresholdMs);
-  }
-
-  shouldRetryBackupSyncWithUcan(error, settings = {}) {
-    const mode = String(settings?.authMode || 'ucan').toLowerCase();
-    if (mode !== 'ucan') return false;
-    const message = String(error?.message || error || '').toLowerCase();
-    return message.includes('401')
-      || message.includes('unauthorized')
-      || message.includes('authentication failed');
-  }
-
-  async ensureBackupSyncUcanFresh(settings = {}, options = {}) {
-    const { force = false } = options;
-    const mode = String(settings?.authMode || 'ucan').toLowerCase();
-    if (mode !== 'ucan') {
-      return { cancelled: false, settings };
-    }
-    if (!force && !this.isBackupSyncUcanExpiringSoon(settings)) {
-      return { cancelled: false, settings };
-    }
-
-    const expiresAt = this.getBackupSyncUcanExpiresAt(settings);
-    if (force) {
-      showSuccess('同步认证失效，正在重新授权');
-    } else if (expiresAt) {
-      showSuccess(`UCAN 即将过期，正在刷新 (${formatDate(expiresAt, 'relative')})`);
-    } else {
-      showSuccess('UCAN 不可用，正在重新授权');
-    }
-
-    return await this.generateBackupSyncUcan();
   }
 
   normalizeUcanResource(value) {
@@ -2641,135 +1607,6 @@ export class SettingsController {
       return normalized !== this.getDefaultUcanResource().toLowerCase();
     }
     return false;
-  }
-
-  async tryBackupSyncAfterUcanAuth(settings) {
-    const enabled = Boolean(settings?.enabled);
-    const endpoint = String(settings?.endpoint || '').trim();
-    if (!enabled || !endpoint) return;
-    try {
-      await this.handleBackupSyncNow();
-    } catch (error) {
-      console.warn('[SettingsController] UCAN 保存后同步失败:', error?.message || error);
-    }
-  }
-
-  async ensureBackupSyncEndpoint() {
-    const input = document.getElementById('backupSyncEndpointInput');
-    const current = String(input?.value || '').trim();
-    if (!current) return '';
-    try {
-      new URL(current);
-    } catch {
-      return current;
-    }
-
-    const resolved = await this.detectBackupSyncEndpoint(current);
-    if (resolved && resolved !== current) {
-      const result = await this.wallet.updateBackupSyncSettings({ endpoint: resolved });
-      if (result?.settings) {
-        this.syncSettings = result.settings;
-        this.renderBackupSyncSettings(result.settings);
-      }
-      if (input) {
-        input.value = resolved;
-      }
-    }
-    return resolved || current;
-  }
-
-  async detectBackupSyncEndpoint(endpoint) {
-    let url;
-    try {
-      url = new URL(endpoint);
-    } catch {
-      return endpoint;
-    }
-
-    const path = url.pathname || '';
-    if (path && path !== '/' && path !== '') {
-      return endpoint;
-    }
-
-    const origin = url.origin;
-    const candidates = ['/', '/dav', '/webdav', '/api'];
-    for (const prefix of candidates) {
-      const ok = await this.probeWebdavPrefix(origin, prefix);
-      if (ok) {
-        return prefix === '/' ? origin : `${origin}${prefix}`;
-      }
-    }
-
-    return endpoint;
-  }
-
-  async probeWebdavPrefix(origin, prefix) {
-    const normalized = prefix === '/' ? '/' : `${prefix.replace(/\/+$/, '')}/`;
-    const url = new URL(normalized, origin);
-    try {
-      const response = await fetch(url.toString(), {
-        method: 'OPTIONS',
-        credentials: 'omit'
-      });
-      if (!response) return false;
-      if (response.status === 404) return false;
-      return true;
-    } catch (error) {
-      return false;
-    }
-  }
-
-  formatBackupSyncError(error, prefix = '同步失败') {
-    const raw = error?.message || '';
-    if (/MKCOL failed:\s*404/i.test(raw)) {
-      return `${prefix}: WebDAV 路径不存在，请检查 WebDAV 地址是否包含正确前缀（如 /dav 或 /api）`;
-    }
-    if (!raw) {
-      return prefix;
-    }
-    return `${prefix}: ${raw}`;
-  }
-
-  async fetchSiweChallenge(endpoint, address) {
-    const url = new URL('/api/v1/public/auth/challenge', endpoint);
-    url.searchParams.set('address', address);
-    const response = await fetch(url.toString(), {
-      method: 'GET',
-      credentials: 'include'
-    });
-    const json = await response.json().catch(() => ({}));
-    if (!response.ok || json.code !== 0) {
-      throw new Error(json.message || 'Challenge failed');
-    }
-    return json.data?.challenge;
-  }
-
-  async fetchSiweVerify(endpoint, address, signature) {
-    const url = new URL('/api/v1/public/auth/verify', endpoint);
-    const response = await fetch(url.toString(), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-      body: JSON.stringify({ address, signature })
-    });
-    const json = await response.json().catch(() => ({}));
-    if (!response.ok || json.code !== 0) {
-      throw new Error(json.message || 'Verify failed');
-    }
-    return json.data || {};
-  }
-
-  async fetchSiweRefresh(endpoint) {
-    const url = new URL('/api/v1/public/auth/refresh', endpoint);
-    const response = await fetch(url.toString(), {
-      method: 'POST',
-      credentials: 'include'
-    });
-    const json = await response.json().catch(() => ({}));
-    if (!response.ok || json.code !== 0) {
-      throw new Error(json.message || 'Refresh failed');
-    }
-    return json.data || {};
   }
 
   async handleRevokeSite(origin, options = {}) {
